test: fix song & chart repository test
