+++ jest.config.all.json
@@ -9,7 +9,7 @@
     ".*\\.(vue)$": "vue-jest",
     "^.+\\.ts$": "ts-jest"
   },
-  "setupFiles": ["jest-canvas-mock"],
+  "setupFiles": ["jest-canvas-mock", "./setupJest.js"],
   "globals": {
     "ts-jest": {
       "tsConfig": "tsconfig.json"
