+++ setupJest.js
@@ -1,0 +1,1 @@
+require('jest-fetch-mock').enableMocks()
